// src/model/account.ts
/*
 * @Author: dbliu shaxunyeman@gmail.com
 * @Date: 2023-01-31 10:18:06
 * @LastEditors: dbliu shaxunyeman@gmail.com
 * @LastEditTime: 2023-01-31 10:20:12
 * @FilePath: /pokeme/src/model/account.ts
 * @Description: 
 */

export interface Account {
    id: string
    mail: string
    name: string
    publicKey: string
}